navigator: relocate enums to their owning stores; drop controllers/constants.ts
CameraMode + BearingMode now live in stores/camera.ts; NavPageKey lives
in stores/nav-sheet.ts. The pagesRequiringMapVisibility Set is no longer
exported — it's a private module-level table feeding a new
NavSheetStore.currentPageRequiresMapVisibility getter (auto-computed via
makeAutoObservable), which AppLayout and MapPaddingStore now consume
directly. AppLayout no longer imports the page-policy table.

maxPortraitSheetCssHeight had a single consumer (AppLayout); inlined
there. controllers/constants.ts is now empty and deleted — the file was
misnamed (nothing in it was controller-specific).

~18 import paths updated across reactions, services, stores, views, and
tests. No behaviour change.
