--- packages/apps/navigator/src/controllers/constants.ts
@@ -1,39 +1,0 @@
-export const enum CameraMode {
-  FOLLOW,
-  FREE,
-}
-
-export const enum BearingMode {
-  MATCH_MAP,
-  NORTH_LOCK,
-}
-
-export const enum NavPageKey {
-  /** shows search bar and destination types */
-  CHOOSE_DESTINATION,
-  /** shows search bar and destination types */
-  SEARCH_ALONG,
-  /** shows instructions for panning/zooming map to choose */
-  CHOOSE_ON_MAP,
-  /** a list of search results that can be routed to */
-  DESTINATIONS,
-  /** a list of routes to a destination */
-  ROUTES,
-  /** step-by-step directions, from an in-progress route */
-  DIRECTIONS_FROM_ROUTE_CONTROLS,
-  /** step-by-step directions, from the `ROUTES` page */
-  DIRECTIONS_FROM_ROUTES_LIST,
-  /** re-order and/or delete waypoints in the active route. */
-  MANAGE_STOPS,
-}
-
-export const navSheetPagesRequiringMapVisibility = new Set<NavPageKey>([
-  NavPageKey.CHOOSE_ON_MAP,
-  NavPageKey.DESTINATIONS,
-  NavPageKey.ROUTES,
-  NavPageKey.DIRECTIONS_FROM_ROUTE_CONTROLS,
-  NavPageKey.DIRECTIONS_FROM_ROUTES_LIST,
-  NavPageKey.MANAGE_STOPS,
-]);
-
-export const maxPortraitSheetCssHeight = '40vh';
