+++ packages/apps/navigator/src/reactions/tests/reactions.test.ts
@@ -8,11 +8,11 @@ import type { IReactionDisposer } from 'mobx';
 import { observable, runInAction } from 'mobx';
 import { vi } from 'vitest';
 import { wireAppReactions } from '..';
-import { CameraMode, NavPageKey } from '../../controllers/constants';
 import type { ChooseOnMapService } from '../../services/choose-on-map';
 import type { MapAdapter } from '../../services/map-adapter';
 import type { RouteRenderer } from '../../services/route-renderer';
-import { NavSheetStoreImpl } from '../../stores/nav-sheet';
+import { CameraMode } from '../../stores/camera';
+import { NavPageKey, NavSheetStoreImpl } from '../../stores/nav-sheet';
 import type { MapPaddingStore } from '../../stores/types';
 import type { AppStore } from './types';
