+++ packages/apps/navigator/src/stores/tests/camera.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { BearingMode, CameraMode } from '../../controllers/constants';
-import { CameraStoreImpl } from '../camera';
+import { BearingMode, CameraMode, CameraStoreImpl } from '../camera';
 
 describe('CameraStoreImpl', () => {
   it('defaults to FOLLOW + MATCH_MAP', () => {
