+++ packages/apps/navigator/src/views/Destinations.tsx
@@ -1,9 +1,9 @@
 import { action } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import { DestinationMarkers } from '../components/DestinationMarkers';
-import { NavPageKey } from '../controllers/constants';
 import { useNavSheetController } from '../services/context';
 import { useNavSheetStore } from '../stores/hooks/use-nav-sheet';
+import { NavPageKey } from '../stores/nav-sheet';
 
 export const Destinations = observer(() => {
   const navSheetStore = useNavSheetStore();
