+++ packages/apps/navigator/src/views/tests/route-controls.test.tsx
@@ -6,8 +6,7 @@ import type { Route } from '@truckermudgeon/navigation/types';
 import { runInAction } from 'mobx';
 import { vi } from 'vitest';
 import type { AppControllerImpl } from '../../controllers/app';
-import { CameraMode } from '../../controllers/constants';
-import { CameraStoreImpl } from '../../stores/camera';
+import { CameraMode, CameraStoreImpl } from '../../stores/camera';
 import { NavSheetStoreImpl } from '../../stores/nav-sheet';
 import { RouteStoreImpl } from '../../stores/route';
 import { RouteControls } from '../RouteControls';
